requirements.txt
langchain==0.1.0
langchain-google-genai==1.0.3
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
httpx==0.27.0
python-dotenv==1.0.0
aiohttp==3.9.1